+++ Api-Java/art/src/main/java/com/dehimik/art/dto/post/PostTagDto.java
@@ -5,6 +5,6 @@ import java.time.Instant;
 
 @Data @NoArgsConstructor @AllArgsConstructor
 public class PostTagDto {
-    private Integer postId;
-    private Integer tagId;
+    private Long postId;
+    private Long tagId;
 }
